masking/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
transformers
numpy
matplotlib
python-dotenv
